refactor: exceptions
