+++ backend/modules/prices/exceptions.py
@@ -8,8 +8,3 @@ class InvalidTimeIntervalException(Exception):
     def __init__(self, reason: str):
         super().__init__(reason)
         self.reason = reason
-
-
-class IllegalDateOrderException(Exception):
-    def __init__(self):
-        super().__init__(f"End date cannot preceed start date")
